Correccion de errores y detalles de accounts

+++ how_to_basic/accounts/views.py
@@ -4,6 +4,7 @@ from django.contrib.auth import login, authenticate
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.forms import AuthenticationForm
 from .models import DatosUsuario
+import os
 # Create your views here.
 
 #Registro de usuario 
@@ -50,7 +51,6 @@ def login_user(request):
 @login_required
 def edit_profile(request):
     usuario=request.user
-    datos_usuario, _ = DatosUsuario.objects.get_or_create(user=usuario)
 
     if request.method == 'POST':
         formulario=UserEditForm(request.POST, request.FILES)
@@ -62,11 +62,8 @@ def edit_profile(request):
                 usuario.last_name=informacion.get('last_name')
             if informacion.get('email'):
                 usuario.email=informacion.get('email')
-            if informacion.get('avatar'):
-                usuario.avatar=informacion.get('avatar')
             if informacion.get('password_1') and informacion.get('password_1') == informacion.get('password_2'):
                 usuario.set_password(informacion.get('password_1'))
-            datos_usuario.save()
             usuario.save()
 
             return render(request, 'index.html', {'usuario':usuario, 'mensaje':'PERFIL EDITADO EXITOSAMENTE'})
@@ -78,7 +75,6 @@ def edit_profile(request):
             'email':usuario.email,
             'first_name': usuario.first_name,
             'last_name': usuario.last_name,
-            'avatar': datos_usuario.avatar
         }
     )
 
@@ -91,19 +87,39 @@ def logout_user(request):
 
 
 #Agregar avatares
-@login_required
+""" @login_required
 def add_avatar(request):
     if request.method == 'POST':
         formulario=AvatarForm(request.POST, request.FILES)
-        if formulario.is_valid():
+        if formulario.is_valid() and len(request.FILES) != 0 :
             avatar_viejo=DatosUsuario.objects.get(user=request.user)
             if(avatar_viejo.avatar):
                 avatar_viejo.delete()
-            avatar_nuevo=DatosUsuario(user=request.user, avatar =formulario.cleaned_data['image'])
+            avatar_nuevo=DatosUsuario(user=request.user, avatar = formulario.cleaned_data['image'])
             avatar_nuevo.save()
             return render(request, 'index.html', {'usuario':request.user, 'mensaje':'AVATAR AGREGADO EXITOSAMENTE'})
     else:
         formulario=AvatarForm()
     return render(request, 'add_avatar.html', {'formulario':formulario, 'usuario':request.user})
+ """
+@login_required
+def add_avatar(request):
+    if request.method == 'POST':
+        formulario = AvatarForm(request.POST, request.FILES)
+        if formulario.is_valid() and len(request.FILES) != 0 :
+            image = request.FILES['image']
+            avatar_v = DatosUsuario.objects.filter(user=request.user.id)
+            if not avatar_v.exists():
+                avatar_n = DatosUsuario(user=request.user, avatar=image)
+            else:
+                avatar_n = avatar_v[0]
+                if len(avatar_n.avatar) > 0:
+                    os.remove(avatar_n.avatar.path)
+                avatar_n.avatar = image
+            avatar_n.save()
+            return render(request, 'index.html', {'usuario':request.user, 'mensaje':'AVATAR AGREGADO EXITOSAMENTE'})
 
+
+    formulario=AvatarForm()
+    return render(request, 'add_avatar.html', {'formulario':formulario, 'usuario':request.user})
